LayoutTests/editing/selection/ios/select-text-by-long-press-with-hardware-keyboard.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true useHardwareKeyboardMode=true ] -->
<html>
<head>
<meta name="viewport" content="width=device-width, initial-scale=1">
<script src="../../../resources/ui-helper.js"></script>
<script src="../../../resources/js-test.js"></script>
<style>
.target {
    font-size: 40px;
    width: 300px;
}

input {
    display: block;
    margin-top: 40px;
    width: 300px;
    height: 80px;
    font-size: 40px;
}
</style>
<script>
jsTestIsAsync = true;
description("Verifies that long-pressing non-editable text while a separate editable element is focused with a hardware (Magic) keyboard attached starts a selection and blurs the focused element (mirrors UIKit TextField behavior: focus and non-editable selection are mutually exclusive).");

addEventListener("load", async () => {
    const input = document.querySelector("input");
    await UIHelper.activateElementAndWaitForInputSession(input);

    window.focusedElementWasInput = document.activeElement === input;
    shouldBeTrue("focusedElementWasInput");

    await UIHelper.longPressElement(document.querySelector(".target"));
    await shouldBecomeEqual("getSelection().toString()", "'Hello'");

    await shouldBecomeEqual("document.activeElement === document.querySelector('input')", "false");
    window.focusedElementBlurred = document.activeElement !== input;
    shouldBeTrue("focusedElementBlurred");

    finishJSTest();
});
</script>
</head>
<body>
    <span class="target">Hello</span>
    <input>
</body>
</html>
